+++ X86-64-correlator/cpu_correlator_2x2_time_sse3.cc
@@ -31,10 +31,10 @@ static unsigned long long calcNrOps(const unsigned nrTimes, const unsigned nrSta
 	for (stat0 = 0; stat0 + 2 <= stat2; stat0 += 2, nrCells ++);
     }
 
-    unsigned long long ops = nrChannels * nrCells * nrTimes * 32L * 4L;
+    unsigned long long ops = (unsigned long long)nrChannels * nrCells * nrTimes * 32L * 4L;
 
-    *bytesLoaded = nrChannels * nrCells * nrTimes * 4L * 4L * 4L;
-    *bytesStored = nrChannels * nrCells * 2L * 4L * 4L;
+    *bytesLoaded = (unsigned long long)nrChannels * nrCells * nrTimes * 4L * 4L * 4L;
+    *bytesStored = (unsigned long long)nrChannels * nrCells * 2L * 4L * 4L;
 
     return ops;
 }
