+++ common/cpu_correlator_1x1.cc
@@ -61,8 +61,8 @@ unsigned long long cpuCorrelator_1x1(const float* __restrict__ samples, float* _
 	}
     }
 
-    *bytesLoaded = nrChannels * nrBaselines * nrTimes * 8L * sizeof(float); // samples
-    *bytesStored = nrChannels * nrBaselines * 8L * sizeof(float); // visibilities
+    *bytesLoaded = (unsigned long long)nrChannels * nrBaselines * nrTimes * 8L * sizeof(float); // samples
+    *bytesStored = (unsigned long long)nrChannels * nrBaselines * 8L * sizeof(float); // visibilities
     
-    return nrChannels * nrBaselines * nrTimes * 16L * 2L; // #operations
+    return (unsigned long long)nrChannels * nrBaselines * nrTimes * 16L * 2L; // #operations
 }
